Webpack config factory, CDN for fontawesome
