+++ src/app.js
@@ -3,7 +3,7 @@ import { render } from 'react-dom'
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import themeDefault from './themeDefault'
 import './styles.scss'
-import 'font-awesome/css/font-awesome.css'
+// import 'font-awesome/css/font-awesome.css'
 import 'flexboxgrid/css/flexboxgrid.css'
 import router from './router'
 import { store } from './redux/configureStore'
